Clean up URLs, add A2CLOUD_SCRIPT_URL
You can now replace http://appleii.ivanx.com/a2cloud/ for script
downloads just as you can with A2SERVER by setting A2CLOUD_SCRIPT_URL in
your shell's environment.  As with A2SERVER, this doesn't affect the
canned files in http://appleii.ivanx.com/a2cloud/files/ because these
aren't really part of the source code.  Certainly it makes sense to
mirror these, but GitHub isn't the place to do that.

Adding everything to the tree was a more manual process than it was for
A2SERVER because there are a lot mroe files in Ivan's tree, and they're
not all used anymore.  More commits were made in the process, but the
end result is that I don't have to go back and spend a bunch of time
undoing my changes for all the URLs that I shouldn't have changed it in
the first place.  Also, it gave me a chance to read through the A2CLOUD
setup script, which was a needed exercise.

Still needed is cleaning up of whitespace to ensure everything is
consistent and adding modelines for vim (and humans who need to know how
to set up editors that aren't vim.)  Those can come as batch commits,
and will follow soon.