setup.txt: Disable everything
Most of this script works fine for jessie, but some of it doesn't and
the script is actually quite long and complex.  Ivan notes he found it
easier to develop it as a cohesive unit that didn't need to worry about
semaphores after doing A2SERVER, and that's obviously going to be
particularly true since he was the only one working on it at the time.

Well, now that's changing we'll want a little more structure.  I won't
be breaking things out into their own files right now, but I'm going to
begin organizing things based on where they should go as I re-enable
them.

Basically, we'll port to jessie block-by-block.