setup.txt: Make 7-bit
There were just a few UTF-8 characters in the changelog section of
setup.txt, so I changed those to ASCII characters.  Sticking with 7-bit
ASCII is probably a good idea for anything that is specifically related
to the Apple // series because there's a possibility it might be viewed
ON such a machine via terminal emulator or something.  Updated Changelog
accordingly.