setup.txt: Remove (comment) A2Pi :(
There are a number of issues regarding A2Pi at the moment:

1. The default A2Pi config or anything resembling it does not presently
   work on Raspberry Pi 3 model B at all due to an upstream bug.

2. The default A2Pi config may not make sense anymore because it was
   really tailored for the Apple II Pi card.  This is a cool device that
   looks like a modified Super Serial Card, but connects your Pi as an
   interal part of your Apple //!  Of course, you can't currently get
   one, it wouldn't be able to power Pi 2/3, etc.

3. Parts of A2Pi like the apple2 user are not updated for Raspbian
   jessie at this time

4. We initially planned one more wheezy release, with me working on
   jessie stuff in the background that was more major and structural in
   nature.  I sort of planned to spin A2Pi out of A2CLOUD during that
   rework and promoting it to its own thing within RasppleII's
   ecosystem.  That way delays in development of either project would
   not hold things back for the other, and because cutting A2CLOUD down
   to size a bit would really make A2CLOUD easier to develop too.

The combination of needing a new release, the sudden dropping of wheezy
support, and the pile of new Pi releases is kinda hurting us a little.
We have a definite regression for A2Pi any way we look at it because of
the Pi3 bug.  :(  And we need to move to Jessie now.

If you're using A2Pi now, this change shouldn't affect you.  A2Pi is
updated through Debian package mechanisms at the moment, so removing the
initial setup from A2CLOUD shouldn't affect existing users on wheezy,
even if they run the next release of the A2CLOUD installer script.  If
you don't have it, this version won't install it for you.

Hopefully we can get back to it soon--I actually use it, so for me it's
a missing feature!  :)