setup.txt: Change most command -v's to hash
This is a change made in A2SERVER already--the choice between the which,
command -v, type, and hash...  The problem is that POSIX defines none of
these so consistently that you can depend on one of them to work.  We
happen to use bash specifically, which gives us an advantage in that
bash's behavior is well-defined.

For us, hash is the best way to find out whether or not a command exists
as it turns out.  So that's what we'll use.