setup.txt: Finish cleanup of GSPort
I'd begun cleaning up GSPort in a branch before we decided to drop A2Pi
and KEGS from the A2CLOUD installer.  Doing that caused the mainline
tree and my branch to differ, so I've been cleaning them both up to work
exactly the same way.  That's done now, only the branch's ### more
searchable comments are different between the two emulator installations
now.

Couldn't just merge the files because there are other changes in the
branch we don't want now, or possibly at all, so I had to do it by
balancing the trees back and forth until the part I was working on was
consistent across both.

I'll probably bring across the searchable comments next, not because
they affect anything, but because they'll assist making sure nothing is
lost as we're working quickly for a Jessie release.