setup.txt: Rewrite XDG/FreeDesktop menu code
The way we've been manipulating the user's LXDE session menu is kind of
not supposed to be done, and certainly it's not the right way to do it.
So what I've done is make sure the correct tool for doing it system-wide
is installed.  It turns out we can actually do it by hand, but because
we're not using a standard category, this method works according to all
of the applicable standards.