setup.txt: Replace scriptURL/binaryURL mechanism
The binaryURL and scriptURL mechanisms sadly don't work properly, and
using it in the first place was kind of a workaround for allowing others
besides Ivan to work on the scripts while we transitioned to doing
things in a way that lent itself to community development.

Here we are over a year later and the scripts are just now undergoing a
rewrite to make them not dependent upon Ivan's webserver.  The mechanism
intended to facilitate that is getting in the way.  It still exists in
some other scripts, but I'm purging it as I go.