New system_ident to set arch-specific vars
Added a new script that sets two variables when sourced, $ras2_os and
$ras2_arch.  These are different than the existing variables we've been
using in a2cloud/a2server and they don't (in and of themselves) presume
Debian.  That said, nothing's likely to work with these scripts except
on Debian, so … yeah.  But the script is designed not to tie you to that
ultimately.

Outputs the variable assignments to stdout unless you pass -q.