Remove cftp entirely, zomg!
Did you know if you're at the top of a directory listing and press up,
it segfaults immediately, leaving your terminal in a modified state
without a visible cursor, forcing you to know how to fix that kind of
broken terminal state with stty, tput, and/or reset?  Yes!

It is at version 0.12 and it was last touched in 2002.  It has no
manpage, no help screen, and the wrong keypress causes it to randomly
die badly.  This is not a "user friendly" program.  If you want ftp with
arrow keys, use lynx!

We're no longer installing this pile of alpha-quality abandonware.
Don't like it?  Make something good or adopt this program and fix its
bugs, then talk to me.  I just can't justify this thing.