Extend the inversion of downloadBinaries
We don't actually download any picopkg files in install_comm_tools
anymore—oysttyer was the only one and we just always pull that from git
and the Readline module from CPAN now.  (In buster we'll only need the
CPAN module if Debian doesn't package it.)

Really the only picopkgs which are optional to download are gsport,
linapple, and nulib.  And since none of those exist at the moment, you
don't even really have much of an "option" about them either.  Yeah, I
know, I'm working on it.