First pass moving emus out of ivan.sh
The install for both LinApple and GSPort is still kind of a mess.  In
LinApple's case that's mostly upstream (tinu.) and GSPort is a
combination of multiple clusterf—it's gonna take some doing.  :)

At some point in the very near future GSPlus will replace GSPort and
that'll clean up some of it.