Move show version code to its own script
…why?  Because show_changes will become the "do you want to install
this?" confirmation at some point.  That's not what it does now, it just
shows you the version for now.  But eventually it will tell you the
version you have installed, compare to the version you are going to
install, show you the changes between here and there, and ask you if you
actually want to proceed with installation.  When that happens it'll be
appropriate to rename the script, but for now it's a placeholder.