Rewrite oysttyer handling, use 2.9.1
We now install oysttyer 2.9.1 instead of old 2.7.2.  While I was at it,
I made the installation prefer the Debian package for oysttyer if you
have it available (in buster and sid), while falling back on local
installations if you don't.

Changed how we remove local installations for Term::ReadKey.  The way we
do it now is less complete, but also less likely to break stuff that we
didn't install maybe.  It doesn't heart anything to leave it behind, but
it requires a manual touch to clean it up.

Finally fixed some missing sudos.  Really the whole installation should
be run with root privs if it's going to run with root privs, but I'm not
ready to begin thinking about that since we don't necessarily know who
the normal user is otherwise, and we still modify the normal user's
account.  (That should change, but we're not in a position to change it
yet.)