New script, install_java!
And what do you think install_java does?  If you just say it's going to
install the latest version of perl, I'm going to hit you.

On Raspbian systems it defaults to oracle-java8-jdk because that's what
Raspbian defaults too _for now_.  Apparently openjdk 9 can keep up with
it and possibly run circles around it, in which case maybe we want to
use default-jre everywhere.  I'm for not using Oracle Java regardless
because … Oracle, but if it's not just as fast, those of you running
armv6-based Pis are going to not be happy if we drop it.