First pass at a Markdown changelog
I went with a Debian changelog format because I figured a2cloud would
probably wind up in Debian someday.  But a Debian changelog is probably
not appropriate for what we want in this file, and I'd prefer something
easier to parse by script anyway.  So I'm transitioning to a markdown
based format described in https://keepachangelog.com/.

Fixed the wrong date on 1.6.7.  November, what?