Clean up gsport-setup a little
We don't do unar picopkgs anymore.  We probably shouldn't even compile
it actually.  The reason why we do is because I don't want to give up on
the idea that a2cloud may run on macOS at some point—but I'm as likely
to just depend on Homebrew for that on the Mac.

Why Homebrew?  Well, I don't like its installation still (same reason I
don't like ours.  Running a command off a website is unwise.  And of
course people do stuff like that all the time, but Homebrew encourages
it and shouldn't!  So again, why Homebrew?

With fink, you build everything from source, but you inherit the Debian
library dependency hell nonetheless. MacPorts solves the dependency hell
and is very robust, but it's also pretty (subjectively) unfriendly to
new users who don't grok UNIX or even know where to find the M to RTFM.

So we come to brew.  It's pretty lightweight, it feels modern.  It's
distributed so that you need not be part of the in-crowd to work with
it, and its Cellar approach provides makes maintenance easy.  When I
first looked at it, it'd been around for years but it had never really
caught on.  Lots of things were not there or just really outdated and
unloved.  I'm glad to say that's changed and the project feels more
mature now.