Yeah, unar's a package, just use Debian's
The Unarchiver for UNIX is available on Debian, Ubuntu, Fedora, and on
macOS under the package name "unar".  It's in Arch Linux as well called
"unarchiver".  What this means is we really have no need whatsoever to
download the source code and compile it, so we won't.

Also made install_archive_tools ignore args it doesn't understand.